requirements.txt
# Browser e Scraping
playwright==1.41.2
beautifulsoup4==4.12.3
lxml==5.1.0

# Processamento de Dados
pandas==2.2.0
numpy==1.26.3
openpyxl==3.1.2
pyarrow==15.0.0

# Utilitários
python-dotenv==1.0.1
requests==2.31.0
aiohttp==3.9.3
asyncio==3.4.3
redis==6.1.0

# Configuration
pydantic==2.6.1
pydantic-settings==2.1.0

# Logging e Monitoramento
prometheus-client==0.19.0

# Testes
pytest==8.0.0
pytest-asyncio==0.23.5
pytest-cov==4.1.0

# Desenvolvimento
black==24.1.1
isort==5.13.2
flake8==7.0.0
mypy==1.8.0

# Mantenha as versões fixas para produção. Revise periodicamente para segurança.